remove useless box implementations
