+++ server/models/sir_h/state.py
@@ -1,6 +1,5 @@
 from models.components.box import BoxSource, BoxTarget
 from models.components.box_dms import BoxDms
-from models.components.box_queue import BoxQueue
 from models.components.box_convolution import BoxConvolution
 from models.components.utils import compute_khi_exp, compute_khi_binom, compute_khi_delay
 from operator import add
