--- server/test/test_box_queue_dms.py
@@ -1,39 +1,0 @@
-import unittest
-from models.components.box_queue_dms import BoxQueueDms
-
-
-class TestBoxDms(unittest.TestCase):
-
-    def test_box_dms1(self):
-        box = BoxQueueDms('DMS-10', 10)
-        box.add(1)
-        self.assertEqual(box.input(), 1)
-        self.assertEqual(box.size(), 0)
-        self.assertEqual(box.output(), 0)
-        box.step()
-        self.assertEqual(box.input(), 0)
-        self.assertEqual(box.size(), 1)
-        self.assertEqual(box.output(), 0)
-
-    def check_input_output(self, box, inputs, r, outputs):
-        for i in range(len(inputs)):
-            box.add(inputs[i])
-            self.assertEqual(box.input(), inputs[i])
-            box.step()
-            self.assertAlmostEqual(box.size(), r[i])
-            self.assertAlmostEqual(box.output(), outputs[i])
-            box.remove(box.output())
-
-    def test_box_dms2(self):
-        box = BoxQueueDms('DMS-10', 10)
-        inputs = [1, 2, 4, 8, 10, 12, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
-        r = [1.000000000, 2.900000000, 6.610000000, 13.949000000, 22.554100000, 32.298690000, 39.068821000, 44.161938900, 47.745745010, 49.971170509, 50.974053458, 50.876648112, 49.788983301, 47.810084971, 45.029076474, 41.526168826, 37.373551944, 33.636196749, 30.272577074, 27.245319367, 24.520787430, 22.068708687, 19.861837819, 17.875654037, 16.088088633, 14.479279770, 13.031351793, 11.728216613, 10.555394952, 9.499855457, 8.549869911,
-             7.694882920, 6.925394628, 6.232855165, 5.609569649, 5.048612684, 4.543751415, 4.089376274, 3.680438647, 3.312394782, 2.981155304, 2.683039773, 2.414735796, 2.173262216, 1.955935995, 1.760342395, 1.584308156, 1.425877340, 1.283289606, 1.154960646, 1.039464581, 0.935518123, 0.841966311, 0.757769680, 0.681992712, 0.613793440, 0.552414096, 0.497172687, 0.447455418, 0.402709876, 0.362438889, 0.326195000, 0.293575500, 0.264217950]
-        outputs = [0.000000000, 0.100000000, 0.290000000, 0.661000000, 1.394900000, 2.255410000, 3.229869000, 3.906882100, 4.416193890, 4.774574501, 4.997117051, 5.097405346, 5.087664811, 4.978898330, 4.781008497, 4.502907647, 4.152616883, 3.737355194, 3.363619675, 3.027257707, 2.724531937, 2.452078743, 2.206870869, 1.986183782, 1.787565404, 1.608808863, 1.447927977, 1.303135179, 1.172821661, 1.055539495, 0.949985546, 0.854986991,
-                   0.769488292, 0.692539463, 0.623285517, 0.560956965, 0.504861268, 0.454375142, 0.408937627, 0.368043865, 0.331239478, 0.298115530, 0.268303977, 0.241473580, 0.217326222, 0.195593599, 0.176034240, 0.158430816, 0.142587734, 0.128328961, 0.115496065, 0.103946458, 0.093551812, 0.084196631, 0.075776968, 0.068199271, 0.061379344, 0.055241410, 0.049717269, 0.044745542, 0.040270988, 0.036243889, 0.032619500, 0.029357550]
-        self.check_input_output(box, inputs, r, outputs)
-
-
-if __name__ == '__main__':
-    unittest.main()
